+++ src/components/TabMenuComponent/containers/HorizontalTabsContainer.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import { SELECTOR_HEIGHT } from '#src/components/TabMenuComponent/constants';
 
 export const HorizontalTabsContainer = styled.div<{ $showUnderline?: boolean }>`
+  position: relative;
   display: flex;
   flex-wrap: nowrap;
   align-items: center;